CimBackendSystem/requirements.txt
alembic==1.13.2
Flask==2.2.2
flask_cors==4.0.1
flask_migrate==4.0.7
flask_sqlalchemy==3.1.1
Requests==2.32.3
schedule==1.2.2

SQLAlchemy==2.0.30
websockets==12.0
Werkzeug==2.2.3
WTForms==3.1.2
xxtea_py==1.0.3
alibabacloud_dyvmsapi20170525==3.2.1
alibabacloud_tea_openapi==0.3.9
alibabacloud_tea_util==0.3.12
coremltools==7.2
matplotlib==3.8.4
numpy==1.26.4
onnx==1.16.1
opencv_python==4.8.1.78
pafy==0.5.5
pandas==2.2.2
Pillow==10.3.0
PyYAML==6.0.1
PyYAML==6.0.1

scipy==1.13.1
seaborn==0.13.2
setuptools==68.0.0
thop==0.1.1.post2209072238
torch==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.4
wandb==0.17.0

